Experiments pnm 12

+++ src/include/gpmon/gpmon.h
@@ -140,7 +140,7 @@ struct gpmon_query_seginfo_t
 	int64				final_rowsout;
 	uint64				sum_cpu_elapsed;
 	uint64				sum_measures_rows_out;
-	int16 				node;
+	uint64 				node;
 };
 
 /* process metrics ... filled in by gpsmon */
